python/api/pyproject.toml
[build-system]
requires = ["hatchling>=1.22.0"]
build-backend = "hatchling.build"

[project]
name = "branchforge-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "gunicorn>=22.0",
  "python-multipart>=0.0.9",
  "pydantic>=2.6",
  "sqlmodel>=0.0.22",
  "redis>=5.0",
  "rq>=1.16",
  "python-dotenv>=1.0",
  "psycopg2-binary>=2.9",
  "alembic>=1.13",
  "PyJWT>=2.8",
  "bcrypt>=4.1",
  "authlib>=1.3",
  "httpx>=0.27",
  "stripe>=8.0",
]

[project.optional-dependencies]
s3 = ["boto3>=1.34"]

[tool.hatch.build.targets.wheel]
packages = ["branchforge_api"]
